flask_app/requirements.txt
cffi==1.13.2
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
six==1.13.0
SQLAlchemy==1.3.11
Werkzeug==0.16.0